Improvements in struct FreqValueSet, SweepReply, SpectranConfigurator, SweepBuilder. Addition of a method to reset sweep in SpectranInterface. Better way to calculate de exact number of samples. Changes in SweepBuilder and SpectranInterface to accelerate the sweep capturing.
Method PushBack() in FreqValueSet was modified to avoid the same frequency to be repeated. A type casting in SweepReply was modified to avoid overflow with big values of frequency. The calculating of number of samples was moved from SpectranConfigurator to SweepBuilder to make sure that the exact number of samples of each sweep is known. The spectran interface's baud rate was increased and the delay to wait from data in sweep capturing loop was deleted, both to accelerate sweep capturing. The folder Debug/ was added to the index to avoid errors.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>