Corrected errors in CurveAdjuster. Improvements in the way that DataLogger saves data. Changes in FreqValueSet. Corrections in the algorithm to calculate front end parameters. Minor changes in SpectranInterface. Improvements in the testbench of the calibration.
An error in the iteration of linear functions and an error in the comparinson of frequency values were corrected in CurveAdjuster. The epsilons were modified, in functions to compare float variables. More overloadings of operator '/' were added in FreqValueSet. The checking of frequency matching was deactived in FreqValueSet due to floating-point rounding errors. Changes in many variables names.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>