Overloading of operators for structure FreqValues and its derived structure Sweep. Simplifying of algorithms in FrontEndCalibrator class via the building of a curve with RBW values vs frequency. Removing of FrontEndParameters structure and saving of those parameters in FreqValues structures. Impovements in the algorithm to calibrate sweeps. Corrections in yaw angle calculation. Improvements in exceptions generated in SpectranConfigurator. Removing of sweep time checking. Correction in main.cpp.
Overloading of sign inversion operator for std::vector<float> container. The algorithm to calibrate sweeps was improved by adding the removing of internal receiver noise from the power measurements, so that they represent just the power coming from the antenna. Method Clear() for TimeData structure. The sweep time parameter do not have to be checked when it is configured because the spectrum analyzer can take a different value taking into account the minimum sweep time.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>