A signal handler for SIGINT and SIGTERM signals was added. Three possible arguments were added to control: on-line curves ploting enabling, front end calibratrion enabling, rfi detection enabling and number of measurements cycles. The data type of frequency values was changed to std::uint_least64. The equality control of frequency values was added. Some operators overloadings for Sweep structure were erased to avoid errors. Changes were made to method FrontEndCalibrator::SaveFrontEndParam() because the frequency data type changed. A method to load default front end parameters was added. A top level header was added and the RFIMS_CART.h was renamed to Basics.h. File FloatComparison.cpp was renamed to Basics.cpp. Methods to set and check Spectran variables were changed.  The sending of commands before the commands VERIFY was disabled. The handling of some errors was changed to ignore them and continue. A method to know if front end parameters have been estimated/loaded was added.
The equality control of frequency values corresponds to operations between FreqValues and/or Sweep structures. Method FreqValues::PushBack() was
changed to compare for exact equality. Function abs() was replaced by fabs() in AntennaPositioner.cpp. The minimum number of satellites of GPS
receiver was changed from 1 to 3. The top level header was added to put there the SignalHandler class which has the handler for signals SIGINT and
SIGTERM. Implementation of function WaitForKey() and sign inversion operator overloading for std::vector<float> were put in Basics.cpp. The way
frequency values are extracted from Spectran replies was changed because of the data type change. Some 'const' qualifiers were added along the
code sources. Correction of sweep time which was adopted by Spectran device, after this was configured, was added.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>