Initial azimuth angle is got from GPS receiver now. The GPS measurements capturing was made more robust. The 'const' modifier was added in each place where was possible. A base path was added in Basics.h to make software parametrised. 'inline' modifiers were removed from header files and some inline functions implementations were moved to the header file to avoid compiler errors. A global enumeration with the Raspberry Pi pins' numbers was added and the specific enumerations with pins numbers were removed from classes. The number of azimuth positions was changed from 8 to 6. Some dumb destructors were added to be called by SignalHandler. New source file with TimeData structure methods' implementations. Several methods were added to TimeData structure. A method to get the mean value of a curve was added to the FreqValues structure. The structure RFI was moved to the Basics.h file. A function to initialize GPIO pins was added. Two functions to turn the RF front end on and off sequentially were added. Two functions were added to process and respond to main's arguments. Global variables were added which define the software behavior taking into account the main()'s arguments. An error in class CurveAdjuster was solved. Deep changes were made in Class DataLogger to support saving in memory of the frequency bands' parameters in a CSV file, compressing of files, deleting of old files and uploading of data.  A class to detect RFI, called RFIDetector, was developed to identify RFI in the last calibrated sweep. A method to plot RFI was added to class RFPloter and some minor changes were made. Some changes were made in class FrontEndCalibrator: a method to build a curve with RBW values vs frequency was added to simplify equations syntax; checking of estimated front end parameters are not ridiculous was added to avoid incorrect sweeps calibration; some errors were solved in methods CalibrateSweep() and the method LoadDefaultParameters() was improved. In class SpectranInterface the method ResetDevice() was renamed to SoftReset() and it was added a method to turn the Spectran device on and off which was called HardReset(). A parameters set name was added to files with frequency bands parameters. The original frequency bands are partitioned into subbands (which has a span less than 200MHz) and the original ones are not discarded now. Some changes were made to the class SignalHandler. Callings to function to turn RF front end off were added in main.cpp.
The global enumeration with Raspberry Pi pins' numbers (Wiring Pi numbers), called 'piPins', was added to unify the use of GPIO pins.
The number of azimuth positions was reduced taking into account the SKA protocol defines the antenna beamwidth between -6dB points and
this had been defined previously between -3dB points, so with a bigger beamwidth less azimuth positions are needed. Overloadings of
operator + were added to the structure Sweep. The aim of class SignalHandler is to respond to signals SIGTERM and SIGINT, destroying all
main()'s objects and then exiting, so all classes must have a destructor, even a foolish one. The new TimeData's methods are: overloadings
of comparison operators and assignement operator, a method to turn back a TimeData object a certain number of days, methods to set only the
date, only the time or both (timestamp); further, the methods to get only the time, only the date or both were renamed. The error of the
class CurveAdjuster was in the method AdjustCurve() and it was caused by the way the linear functions (or just lines) were swept. The
method to save estimated front end parameters was moved to DataLogger class. In class DataLogger: method to save frequency bands'
parameters into memory in a CSV file was added to upload these with that format; method DeleteOldFiles() removes the files older than a
month; method CompressLastFiles() was added to archive and compress the files which will be uploaded; and method UploadData() calls a
python script to send the RF data to a remote server. In the class SignalHandler were added some pointers to objects which must be
destroyed and it was added a call to the function to turn RF front end off sequentially.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>