Improving comments to generate the Doxygen manual. Cleaning the code. Minor changes were made in the code. The directories tree was changed. A folder named scripts/ was added to put there the scripts related with this software. A user-defined makefile was added.
The python script "client.py", which is called by this software, was added to the folder scripts/. The shell script
"rfims-cart-boot", which executes the rfims-cart software at the computer booting, was added to the folder scripts/.
A folder named src/ was added and the source and header files were moved there. A folder named doc/ was added, where
the generated documents will be stored, like the doxygen manual. The makefile was added to the root directory and this
is intended to build the software from a terminal. The makefile will store the objects in a folder named obj/ and the
executable file (binary file) will be stored with the name "rfims-cart" and in a folder named bin/.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>