The GPSInterface class was modified to simplify the capture of the sensors' data with the streaming disabled and to avoid errors when the checksum is wrong.
It was added several methods to update the values of certain attributes, taking into account only the corresponding data replies and ignoring the other ones. These methods just update the attributes once, so they are the opposite to the streaming where the attributes are constantly updated. The streaming will be enabled later. The main function was modified following these changes.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>