The GPSInterface class was improved for this one to support the data streaming of the Aaronia GPS Logger. A new method to enable the streaming was added. The 'Update' methods (streaming disabled) were improved to avoid errors when the streaming is enabled. It was added a new 'Update' method to just update the number of satellites. Several methods called 'New...' were added to know if there are new data when the data streaming is enabled. Two methods were added to get the streaming data rate and to know if the streaming is enabled or not. A testbench was added to test the Aaronia GPS Logger; it was put in a new folder, test/. The makefile was modified to allow the compilation of the GPS Logger's testbench.
The streaming data rate of the sensors data was reduced to 2 data/s. The storing of the sensors data in files for debugging purposes was removed. Several flags were added to know if there are new data, and the methods 'Extract', 'Calculate' and 'Get' were modified to manipulate these flags. To process the data when the streaming is enabled, it was added a thread which get and extract the data from each reply of the GPS Logger; this thread is created when the program enables the streaming and it is cancelled when the program disables the streaming. The method GPSInterface::Read was improved to overcome errors.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>