A testbench for the Spectran device was created, it is intended to capture sweeps with the spectrum analyzer connected to a RF front end. The testbench receives as arguments the number of files where the sweeps must be stored and the number of sweeps per file. Also it receives an argument which indicates if the front end calibration must be performed or not and another argument which determines if the plotting must be performed or not. The global variables were moved from main.cpp to TopLevel.cpp to avoid to add them in the testbenches. Function PrintHelp() was improved. Minor changes were made in file main.cpp
The classes DataLogger and AntennaPositioner were modified to have internals variables to store the number of azimuth positions and the number of sweeps and, in that way, avoid to use a global external variable. A method to set these internal variables was added to each class. The variable 'numOfAzimPos' was moved to TopLevel.cpp because now it is just needed in the functions main() and ProcessMainArguments(). It was necessary to declare global variables in TopLevel.cpp.

Signed-off-by: MauroDiamantino <maurogabriel.54@gmail.com>